% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAmps.R
\name{calcAmps}
\alias{calcAmps}
\title{reading a list of data}
\usage{
calcAmps(lamp)
}
\arguments{
\item{lamp}{a list of data that plot by the cycling data}
}
\value{
a list of data with calculated amplitude
}
\description{
reading a list of data of cycling lamp data, and return a list with calculated amplitude
}
