% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAmps2.R
\name{calcAmps2}
\alias{calcAmps2}
\title{reading a list of data}
\usage{
calcAmps2(lamp)
}
\arguments{
\item{lamp}{a list of data that plot by the cycling data}
}
\value{
a list of data with calculated amplitude
}
\description{
reading a list of data of cycling lamp data, and return a list with calculated amplitude
the backup plan from calcAmps, delet the nm520 and specific the rows and cols etc.
}
