% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAmpsGeneric.R
\name{calcAmpsGeneric}
\alias{calcAmpsGeneric}
\title{reading a list}
\usage{
calcAmpsGeneric(lamp, ...)
}
\arguments{
\item{lamp}{a list that have the information of the lamp}

\item{...}{any additional parameter that needed to put in}
}
\value{
A list containing the amplification results and a summary of positive amplifications.
}
\description{
reading a list and return a list that content the calculated amplitude in generic version
}
