% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAmpsSNP.R
\name{calcAmpsSNP}
\alias{calcAmpsSNP}
\title{Read lamp data}
\usage{
calcAmpsSNP(lamp)
}
\arguments{
\item{lamp}{a lamp list that contain lamp data}
}
\value{
a list with calculated amplitude an position
}
\description{
Reading a lamp data and return a list with calculated amplitude and position
}
