% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCt.R
\name{calcCt}
\alias{calcCt}
\title{Calculate the cycle threshold (Ct) values for qPCR data}
\usage{
calcCt(
  fluor,
  well,
  meta = NULL,
  threshold = 1e+05,
  timePerStep = 0.5,
  digits = 1,
  maxTime = 60
)
}
\arguments{
\item{fluor}{a vector that indicate the fluorescence data}

\item{well}{a vector that indicate the well identifiers}

\item{meta}{An optional data frame content metadata for each well}

\item{threshold}{An int that indicate the threshold of the fluorescence, initial with 100000}

\item{timePerStep}{A double that indicate the time take per step, initial with 0.5}

\item{digits}{An int that indicate the number of decimal of the calculated threshold round up to, initial with 1}
}
\value{
A data frame that content Ct information and wells identifiers etc.
}
\description{
This function calculates the Ct values for quantitative PCR (qPCR) data based on fluorescence readings.
}
