% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkAmps.R
\name{checkAmps}
\alias{checkAmps}
\title{read list of data from readStepone}
\usage{
checkAmps(
  lamp,
  melt,
  fl,
  minAmp = 2500,
  meltTempNum = 35,
  meltTempNum2 = length(unique(melt$temp)),
  minFoldIncrease = 3,
  minMeltDiff = 0.5,
  baselineTime = 1,
  isTemp = FALSE
)
}
\arguments{
\item{lamp}{a list of data provided by lamp file}

\item{melt}{a list of data provided by melt file}

\item{minAmp}{an int number indicate the minimum amplitude, initial with 2500mn}

\item{meltTempNum}{an int number indicate the melting temperature, initial with 35 degrees}

\item{meltTempNum2}{an int number **}

\item{minFoldIncrease}{an int number, initial with 3}

\item{minMeltDiff}{a number that indicate the minimum difference between each melt data, initial with 0.5}

\item{baselineTime}{a number that indicate the staring base time is}

\item{isTemp}{a Boolean value **}
}
\description{
yhid read the data list and to check the amplitude from the previous functions
}
