% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colRowSelection.R
\name{colSelection}
\alias{colSelection}
\title{No parameter needed}
\usage{
colSelection(numC = 12)
}
\value{
a squeal that tells the range of the column
}
\description{
With no parameter, the function will promote the user to input the col information
}
