% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condenseSamples.R
\name{condenseSamples}
\alias{condenseSamples}
\title{Condense excel file data}
\usage{
condenseSamples(files, pattern48 = "^48", rows = 1:8, cols = 1:6)
}
\arguments{
\item{files}{An excel file that have the raw data of lamp}

\item{pattern48}{an expression pattern that to match sheet name initial with '^48'}

\item{rows}{the selection of rows to read in the file, initial with 1:8}

\item{cols}{the selection of cols to read in the file, initial with 1:6}
}
\value{
A data frame containing condensed sample data with the corresponding file and sheet names.
}
\description{
Reading excel file and used it to output data frame
}
