% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSamples.R
\name{findSamples}
\alias{findSamples}
\title{Finding sample from an excel file}
\usage{
findSamples(samples, files, outFile = NULL)
}
\arguments{
\item{samples}{A string that indicate the samples name}

\item{files}{An excel file that content lamp raw data}

\item{outFile}{a}
}
\description{
Using sample name to find matching samples in an excel file that content raw data 
and return organized sample information in list
}
