% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findWellLine.R
\name{findWellLine}
\alias{findWellLine}
\title{read an excel file}
\usage{
findWellLine(xls)
}
\arguments{
\item{xls}{excel file of which the raw data are}
}
\description{
read an excel file and promote the massage of the wells information
}
