% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFileName.R
\name{GetFileName}
\alias{GetFileName}
\title{No parameter needed}
\usage{
GetFileName()
}
\value{
a string
}
\description{
With no parameter, the function will promote the user to input a file name that will later be used
}
