% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overAll.R
\name{overAllColBase}
\alias{overAllColBase}
\title{A xlsx file containing lamp cycle raw data}
\usage{
overAllColBase(file1)
}
\value{
a squeal that tells the range of the column
}
\description{
This is an overall funtion that call out others to output the pdf files and a log file (recording the )
}
