% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLamp.R
\name{plotLamp}
\alias{plotLamp}
\title{read a list}
\usage{
plotLamp(lamp, rowStandardize = FALSE)
}
\arguments{
\item{lamp}{a list that contains raw data}

\item{rowStandardize}{a bool that indicate if the data is in the correct form or not}
}
\description{
reads a list of raw data and plot it into the correct form
}
