% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPats.R
\name{plotPats}
\alias{plotPats}
\title{Reading lamp data}
\usage{
plotPats(
  lamp,
  pos = NULL,
  primers = c(E1 = "520nm", STATH = "587nm", As1e = "623nm", Penn = "682nm"),
  plotMelts = !is.null(lamp$melt),
  xVar = "timeMin",
  xlab = "Time (m)",
  minorPos = NULL,
  nPages = 1
)
}
\arguments{
\item{lamp}{a list of lamp data}

\item{pos}{position initial with NULL}

\item{primers}{a list that have the primers used in the lamp, initial with 'E1'='520nm','STATH'='587nm','As1e'='623nm','Penn'='682nm'}

\item{plotMelts}{a bool value that check if the lamp list is empty or not}

\item{xVar}{A string that label the y tag, initial with timeMin}

\item{xlab}{a string that label the x tag, initial with Time(m)}

\item{minorPos}{not sure what that is}

\item{nPage}{an int value}
}
\description{
Reading a list that have lamp data and plot the information into table format
}
