% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSummary.R
\name{plotSummary}
\alias{plotSummary}
\title{read list of data}
\usage{
plotSummary(
  melt,
  xCol,
  xlab,
  fls,
  rowIds,
  colIds,
  cols,
  showLegend = TRUE,
  plotCol = "row",
  lineCol = "col",
  legendInset = c(-0.16, -0.24),
  ncol = 2,
  sameY = FALSE,
  showMain = TRUE,
  yScale = 1000,
  ylab = "",
  ...
)
}
\arguments{
\item{xCol}{description}

\item{xlab}{description}

\item{fls}{description}

\item{cols}{description}

\item{showLegend}{a bool value}

\item{plotCol}{a string to indicate the name tag of the rows}

\item{lineCol}{a string to indicate the name tag of the cols}

\item{legendInset}{description}

\item{ncol}{a number that indicate the number of cols}

\item{sameY}{description}

\item{showMain}{a bool value}

\item{yScale}{description}

\item{ylab}{description}

\item{lamp}{description}

\item{legendArgs}{description}

\item{indicate}{description}

\item{indicateYellow}{description}

\item{mainExtra}{description}

\item{logXY}{description}

\item{ylims}{description}

\item{lwd}{description}

\item{plotPoints}{description}

\item{addYFirst}{description}

\item{cex.lab}{description}

\item{yMaxs}{description}

\item{yScaleFirst}{description}

\item{extraCmds}{description}
}
\description{
plot the data in the correct form and be able to use in later
}
