% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLamp.R
\name{readLamp}
\alias{readLamp}
\title{read NCBI names file}
\usage{
readLamp(
  rawFile,
  meltFile = rawFile,
  meltStart = 121,
  timePerCyle = 30,
  skipCycles = 0,
  skip = 21,
  sep = ",",
  correctCycles = FALSE,
  rawDataTab = "Raw Data",
  meltRawTab = "Melt Curve Raw Data"
)
}
\arguments{
\item{rawFile}{a string giving the path to an NCBI name file to read from (both gzipped or uncompressed files are ok)}

\item{meltFile}{a string giving the path to a NCBI name file, work the same as rawFile, this is a backup if there are second file reading}

\item{meltStart}{a number indicate the staring cycle of the melt curve.}

\item{timePerCyle}{a number indicate the seconds per cycle}

\item{skipCycles}{a number indicate where to read the cycles}

\item{skip}{description}

\item{sep}{a string indicate how to separate the data from}

\item{rawDataTab}{the name of the tab Excel sheet}

\item{meltRawTab}{the name of the tab Excel sheet of melt curve}

\item{correctCycle}{a bool value, that set to tell if the file is in right version}
}
\value{
a list of lamp and melt amplitude table
}
\description{
yhid function reads from excel files generated by the ABI StepOnePlus. read from the (8) line and initial the cycles, return a list of lamp and melt amplitude
}
