% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readStepOne.R
\name{readStepOne}
\alias{readStepOne}
\title{read NCBI names file}
\usage{
readStepOne(
  rawFile,
  rawDataTab = "Raw Data",
  skip = 7,
  cycles = 120,
  minPerCycle = 0.5
)
}
\arguments{
\item{rawFile}{string giving the path to an NCBI name file to read from (both gzipped or uncompressed files are ok)}

\item{rawDataTab}{the name of the tab Excel sheet}

\item{skip}{a number that tells where to star reading the file}

\item{cycles}{a int that set the range of the cycles of the lamp and melt curve}

\item{minPerCycle}{a number that says how many minutes per cycle}
}
\value{
a list of lamp and melt amplitude table
}
\description{
yhid function reads from excel files generated by the ABI StepOnePlus. read from the (8) line and initial the cycles, return a list of lamp and melt amplitude
}
