% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readXls.R
\name{readXls}
\alias{readXls}
\title{read excel file}
\usage{
readXls(xls, isQS6 = FALSE, extraTemps = c(), nCycle = 200)
}
\arguments{
\item{xls}{an excel file that have the raw data}

\item{isQS6}{a Boolean value indicate if the data come from QuantStudio 6}

\item{extraTemps}{description ***}

\item{nCycle}{a int number indicate how many cycle had in the data}
}
\value{
a formatted list **
}
\description{
reading excel file and return with formatted lamp data list
}
