% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colRowSelection.R
\name{rowSelection}
\alias{rowSelection}
\title{No parameter needed}
\usage{
rowSelection(numR = 8)
}
\value{
a squeal that tells the range of the rows
}
\description{
With no parameter, the function will promote the user to input the row information
}
